pyproject.toml
[project]
name = "agent-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.7",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "raw-agents==0.1.2",
    "uvicorn>=0.40.0",
    "yoyo-migrations>=9.0.0",
]
